Finalizando a autenticação dos usuários e rotas de task

+++ src/controller/TaskController.ts
@@ -1,6 +1,7 @@
 import {prisma} from "../config/PrismaConfig"
 import { Request, Response, NextFunction } from "express"
 import { AppError } from "../errors/AppError"
+import id from "zod/v4/locales/id.js"
 
 export class TaskController {
     async list(request: Request, response: Response) {
@@ -8,24 +9,38 @@ export class TaskController {
         const limit = Number(request.query.limit) || 10
         const skip = (page - 1) * limit
 
+        const userId = request.user?.id
+
+        if(!userId) {
+            return response.status(401).json({message: "Usuário não autenticado"})
+        }
+
         const tasks = await prisma.task.findMany({
+            where: {userId: userId},
             skip,
             take: limit,
             orderBy: {createdAt: "asc"}
-        })
+            })
 
-        const total = await prisma.task.count()
+        const total = await prisma.task.count({where: {userId: userId}})
+
         const totalPages = Math.ceil(total / limit)
 
-        return response.status(200).json({message: "Lista paginada de tasks", page, totalPages, total, tasks})
+        return response.status(200).json({message: "Lista paginada de tasks", page, totalPages, total, tasks, user: request.user})
     }
 
     async create(request: Request, response: Response) {
         const {title} = request.body
+        const userId = request.user?.id
+
+        if(!userId) {
+            throw new AppError("Usuário não autenticado ou token inválido", 401)
+        }
 
         const newTask = await prisma.task.create({
             data: {
-                title
+                title,
+                userId: userId
             }
         })
 
@@ -34,9 +49,10 @@ export class TaskController {
 
     async find(request: Request, response: Response, next: NextFunction) {
         const {id} = request.params
+        const userId = request.user?.id
 
         try {
-            const task = await prisma.task.findUnique({where: {id}})
+            const task = await prisma.task.findUnique({where: {id, userId: userId}})
 
             if(!task) {
                 throw new AppError("Task não encontrada")
@@ -52,10 +68,11 @@ export class TaskController {
 
     async deleteTask(request: Request, response: Response, next: NextFunction) {
         const {id} = request.params
+        const userId = request.user?.id
 
         try {
 
-            const taskForDelete = await prisma.task.delete({where: {id}})
+            const taskForDelete = await prisma.task.delete({where: {id, userId}})
 
              return response.status(200).json({ messsage: "Task deletada com sucesso", taskForDelete})
             
@@ -69,9 +86,10 @@ export class TaskController {
         const {id} = request.params
 
         const {...data} = request.body 
+        const userId = request.user?.id
 
         try {
-            const taskForUpdate = await prisma.task.update({where: {id}, data: data})
+            const taskForUpdate = await prisma.task.update({where: {id, userId}, data: data})
 
             return response.status(200).json({message: "Task atualizada com sucesso", taskForUpdate})
